move lazy doc to separate extension
