+++ docs/conf.py
@@ -161,152 +161,7 @@ if tags.has("run_apidoc"):
         shutil.copyfile(doc_dir / "api/api_index.rst", api_index_target)
     print("breathe apidoc completed")
 else:
-    #  if not api_index_target.exists():
-    #  shutil.copyfile(doc_dir / "api/api_stub.rst", api_index_target)
-    pass
-
-# Find all roles pointing at code
-# @TODO: Member is difficult, unscoped enums don't work
-roles = [
-    "class",
-    "struct",
-    "type",
-]
-
-role_names = {"class": "Classes", "struct": "Structs", "type": "Types", "enum": "Enums"}
-
-directives = {
-    "class": "doxygenclass",
-    "struct": "doxygenstruct",
-    "type": "doxygentypedef",
-    "func": "doxygenfunction",
-    "enum": "doxygenenum",
-}
-
-role_instances = {k: set() for k in roles}
-
-role_instances["type"] |= {
-    "Acts::ActsScalar",
-    "Acts::ActsVector",
-    "Acts::ActsMatrix",
-    "Acts::ActsSquareMatrix",
-    "Acts::SquareMatrix2",
-    "Acts::SquareMatrix3",
-    "Acts::SquareMatrix4",
-    "Acts::BoundMatrix",
-    "Acts::BoundSquareMatrix",
-    "Acts::Vector2",
-    "Acts::Vector3",
-    "Acts::Vector4",
-    "Acts::BoundVector",
-    "Acts::BoundTrackParameters",
-    "Acts::Transform2",
-    "Acts::Transform3",
-    "Acts::AngleAxis3",
-    "Acts::RotationMatrix2",
-    "Acts::RotationMatrix3",
-    "Acts::Translation2",
-    "Acts::Translation3",
-    "Acts::GeometryContext",
-    "Acts::FreeVector",
-    "Acts::FreeMatrix",
-    "Acts::SurfaceVector",
-    "Acts::Intersection3D",
-    "Acts::OrientedSurface",
-    "Acts::OrientedSurfaces",
-    "Acts::BoundToFreeMatrix",
-    "Acts::FreeToBoundMatrix",
-    "Acts::FreeSquareMatrix",
-    "Acts::FreeToPathMatrix",
-}
-
-role_instances["struct"] |= {
-    "Acts::DenseStepperPropagatorOptions",
-    "Acts::Experimental::DetectorNavigator::State",
-    "Acts::Geant4PhysicalVolumeSelectors::AllSelector",
-    "Acts::Geant4PhysicalVolumeSelectors::NameSelector",
-}
-
-role_instances["class"] |= {
-    "Acts::BinningData",
-    "Acts::Direction",
-    "Acts::ConstrainedStep",
-    "Acts::IAxis",
-    "Acts::SeedFilter",
-    "Acts::BoundaryCheck",
-    "Acts::ConeVolumeBounds",
-    "Acts::CuboidVolumeBounds",
-    "Acts::CylinderVolumeBounds",
-    "Acts::CutoutCylinderVolumeBounds",
-    "Acts::GenericCuboidVolumeBounds",
-    "Acts::TrapezoidVolumeBounds",
-    "Acts::GeometryObject",
-    "Acts::TrackContainer",
-    "Acts::ConeLayer",
-    "Acts::CylinderLayer",
-    "Acts::IdentifiedDetectorElement",
-    "Acts::DiscLayer",
-    "Acts::PlaneLayer",
-    "Acts::NullBField",
-    "Acts::DiscBounds",
-    "Acts::PlanarBounds",
-    "Acts::AbstractVolume",
-    "Acts::AnnulusBounds",
-    "Acts::DiamondBounds",
-    "Acts::ConvexPolygonBounds",
-    "Acts::ConvexPolygonBoundsBase",
-    "Acts::Logging::LevelOutputDecorator",
-    "Acts::Logging::NamedOutputDecorator",
-    "Acts::Logging::ThreadOutputDecorator",
-    "Acts::Logging::TimedOutputDecorator",
-    "Acts::Logging::DefaultFilterPolicy",
-    "Acts::Logging::DefaultPrintPolicy",
-}
-
-role_instances["enum"] = {
-    "Acts::BinningValue",
-    "Acts::BinningType",
-    "Acts::BinningValue",
-    "Acts::BoundIndices",
-    "Acts::FreeIndices",
-}
-
-role_ex = re.compile(r"[{:](" + "|".join(roles) + r")[}:]`(.+?)`")
-
-
-def process_roles(file: Path) -> List[Tuple[str, str]]:
-    text = file.read_text()
-    return [m.groups() for m in role_ex.finditer(text)]
-
-
-for dirpath, _, filenames in os.walk(doc_dir):
-    dirpath = Path(dirpath)
-    for file in filenames:
-        file = dirpath / file
-        if file.suffix not in (".rst", ".md"):
-            continue
-        for role, arg in process_roles(file):
-            role_instances[role].add(arg)
-
-# add members to their parents
-
-api_preamble = """
-
-"""
-
-with api_index_target.open("w") as fh:
-    fh.write("# API Reference\n\n")
-    fh.write(api_preamble)
-    for role, instances in sorted(role_instances.items(), key=lambda x: x[0]):
-        fh.write(f"## {role_names[role]}\n")
-        for instance in sorted(instances):
-            fh.write(
-                f"""
-:::{{{directives[role]}}} {instance}
-:::
-"""
-            )
-        fh.write("\n")
+    extensions += ["lazy_autodoc"]
 
 
 # -- Markdown bridge setup hook (must come last, not sure why) ----------------
